requirements.txt
python-dotenv
pvporcupine
pvrecorder
pyaudio
azure-cognitiveservices-speech
openai
serial
pyserial
requests